feat(utils): add clamp helper and adopt across packages
