+++ packages/utils/src/number.ts
@@ -3,3 +3,9 @@ export function roundTo(value: number, precision = 2) {
 
     return Math.round(value * x) / x;
 }
+
+export const clamp = (
+    value: number,
+    min = Number.NEGATIVE_INFINITY,
+    max = Number.POSITIVE_INFINITY,
+) => Math.min(Math.max(value, min), max);
